+++ CHANGES.md
@@ -1,3 +1,9 @@
+Unreleased
+-----------
+
+- h2-mirage: adapt to Conduit v4 and remove `H2_mirage.Server_with_conduit`
+  ([#154](https://github.com/anmonteiro/ocaml-h2/pull/154))
+
 0.8.0 2021-04-11
 ---------------